move dh/gen/primes to lib
