+++ lib/dh.js
