+++ lib/generatePrime.js
